[react-devtools] AI chat: provider architecture v2 (P0 core refactor)
Restructures the ad-hoc provider presets into the opencode-style split so
adding providers and auth methods is cheap, with no behavior change for
existing providers.

- providers.js is now a data catalog: each entry declares its wire dialect
  and auth method ({id, label, baseUrl, wire, auth, models}). Custom is
  removed (an OpenAI-compatible endpoint is reached by editing a provider's
  base URL); OpenAI is added.
- providerRuntime.js is the auth loader: resolveRequest(config) turns stored
  config into a ResolvedRequest {wire, baseUrl, model, headers} or a
  user-facing error. This is the seam where P1's Codex subscription auth
  (token import + refresh + ChatGPT-Account-Id header) will plug in.
- wire/openaiChat.js is the chat-completions adapter, extracted verbatim
  from the old client and parameterized by the resolved request.
- client.js is now a thin dispatcher: resolve then dispatch by wire.

Settings render from the catalog: the API key field shows only for api-key
providers (hidden for local Ollama). Verified end to end via local Ollama —
streaming, tool calling, and grounded answers all work through the new
dispatch path. (DevTools test suite: 607 real tests pass; the one failing
suite is the extension-build test, which needs disk space unavailable in
this environment.)
